+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html) with
 the exception that the versions 0.*.* may have breaking changes in minor versions.
 
+## [0.5.8]
+### Added
+- Add a check to prevent adding duplicate balance assertions in the `Journal` class.
+- Add i18n for file headers
+
+### Removed
+- Removed unused functions in the `report` module.
+
 ## [0.5.7]
 ### Fixed
 - `Txn` constructor now checks that the dates of the postings are the same.